Fixed progress calculations

+++ src/main/java/com/redhat/topicindex/extras/client/local/App.java
@@ -149,8 +149,8 @@ public class App
 		/* There are files to process, process them and then move to the next file */
 		else
 		{
-			final float blockLevelProgress = blockIndex / this.imageUploadBlocks.size();
-			final float fileLevelProgress = fileIndex / data.getUpload().getFiles().getLength() / this.imageUploadBlocks.size();
+			final float blockLevelProgress = (float)blockIndex / this.imageUploadBlocks.size();
+			final float fileLevelProgress = (float)fileIndex / data.getUpload().getFiles().getLength() / this.imageUploadBlocks.size();
 			
 			progress.setPercentDone((int) ((blockLevelProgress + fileLevelProgress)  * 100));
